split db media services
